pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weather-cli"
version = "0.1.0"
description = "A Python CLI tool that fetches and displays weather information"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Developer", email = "dev@example.com"}]
keywords = ["weather", "cli", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.25.0",
    "requests-cache>=1.0.0",
    "click>=8.0.0",
    "tabulate>=0.8.0",
    "pyyaml>=6.0",
    "pytest>=6.0.0",
]

[project.scripts]
weather-cli = "weather_cli.weather:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]